PEP-514 support via winreg (#1310)
* pep-514 implementation

* reorganize interpreter

* report failures

* report failures
