+++ docs/changelog/1306.bugfix.rst
@@ -1,0 +1,1 @@
+Turns out the output of the ``py -0p`` is not stable yet and varies depending on various edge cases. Instead now we read the interpreter values directly from registry via `PEP-514 <https://www.python.org/dev/peps/pep-0514>`_ - by :user:`gaborbernat`.
