+++ tox.ini
@@ -95,7 +95,7 @@ commands = echo {posargs}
 [flake8]
 max-complexity = 22
 max-line-length = 99
-ignore = E203, W503, C901, E402
+ignore = E203, W503, C901, E402, B011
 
 [pep8]
 max-line-length = 99
@@ -129,6 +129,9 @@ rsyncdirs = tests tox
 looponfailroots = tox tests
 testpaths = tests
 xfail_strict = True
+markers =
+    git
+    network
 
 [isort]
 multi_line_output = 3
@@ -136,7 +139,7 @@ include_trailing_comma = True
 force_grid_wrap = 0
 line_length = 99
 known_first_party = tox,tests
-known_third_party = apiclient,docutils,filelock,flaky,freezegun,git,httplib2,oauth2client,packaging,pathlib2,pkg_resources,pluggy,py,pytest,setuptools,six,sphinx,toml
+known_third_party = apiclient,docutils,filelock,flaky,freezegun,git,httplib2,oauth2client,packaging,pathlib2,pkg_resources,pluggy,py,pytest,setuptools,six,sphinx,toml,winreg
 
 [testenv:release]
 description = do a release, required posarg of the version number
